requirements_dev.txt
pip==9.0.1
bumpversion==0.5.3
wheel==0.29.0
watchdog==0.8.3
flake8==3.4.1
tox==2.8.1
coverage==4.5.1
Sphinx==1.7.0
cryptography==2.0.3
PyYAML==3.12
nose==1.3.7
dateutils==0.6.6
